upload explorer()

+++ pop/pop_360.py
@@ -67,22 +67,21 @@ def bingdu():
 
 
 def explorer():
-    logger.info('*******explorer stopped*********')
-    type(Key.F11)
-    wait(0.1)
-    click(Pattern("explorer.png").targetOffset(0, 100))
-    wait(0.1)
-
-
-def windows():
-    try:
-        subprocess.call('explorer', shell=True)
-    except Exception as e:
-        logger.info(e)
-    try:
-        subprocess.call('powershell.exe Stop-Process -name explorer', shell=True)
-    except Exception as e:
-        logger.info(e)
+    if exists("explorer.png", 1):
+        logger.info('*******explorer stopped*********')
+        type(Key.F11)
+        wait(0.1)
+        click(Pattern("explorer.png").targetOffset(0, 100))
+        wait(0.1)
+    if not exists("windows.png", 2):
+        try:
+            subprocess.call('explorer', shell=True)
+        except Exception as e:
+            logger.info(e)
+        try:
+            subprocess.call('powershell.exe Stop-Process -name explorer', shell=True)
+        except Exception as e:
+            logger.info(e)
 
 
 def UI():
@@ -90,10 +89,6 @@ def UI():
     t.setDaemon(True)
     t.start()
     try:
-        if exists("explorer.png", 2):
-            explorer()
-        if not exists("windows.png", 2):
-            windows()
         if exists("install.png", 1):
             install()
         if exists("fil.png", 1):
@@ -204,6 +199,7 @@ def pb():
 
 
 def b4hand(project, package, updc, p_list):
+    explorer()
     UI()
     inst(package)
     setTime = [setTime2M]
