+++ pop/pop_qq.py
@@ -38,22 +38,21 @@ def yes():
 
 
 def explorer():
-    logger.info('*******explorer stopped*********')
-    type(Key.F11)
-    wait(0.1)
-    click(Pattern("explorer.png").targetOffset(0, 100))
-    wait(0.1)
-
-
-def windows():
-    try:
-        subprocess.call('explorer', shell=True)
-    except Exception as e:
-        logger.info(e)
-    try:
-        subprocess.call('powershell.exe Stop-Process -name explorer', shell=True)
-    except Exception as e:
-        logger.info(e)
+    if exists("explorer.png", 1):
+        logger.info('*******explorer stopped*********')
+        type(Key.F11)
+        wait(0.1)
+        click(Pattern("explorer.png").targetOffset(0, 100))
+        wait(0.1)
+    if not exists("windows.png", 2):
+        try:
+            subprocess.call('explorer', shell=True)
+        except Exception as e:
+            logger.info(e)
+        try:
+            subprocess.call('powershell.exe Stop-Process -name explorer', shell=True)
+        except Exception as e:
+            logger.info(e)
 
 
 def UI():
@@ -167,6 +166,7 @@ def pb():
 
 
 def b4hand(project, package, updc, p_list):
+    explorer()
     UI()
     inst(package)
     setTime = [setTime2M]
